chore : change file structure
