+++ src/components/Image/ImageCarousel/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Modal from 'components/Modal';
-import PrevNextButton from 'components/PrevNextButton';
+import Modal from 'components/ui/Modal';
+import PrevNextButton from 'components/ui/PrevNextButton';
 import { useImageSlide } from 'hooks';
 import { baseModalStyle } from 'common/style/baseStyle';
 import styled from '@emotion/styled';
