+++ src/components/Image/index.ts
@@ -1,0 +1,3 @@
+export { default as ImageCarousel } from './ImageCarousel';
+export { default as ImagePreview } from './ImagePreview';
+export { default as ImageSlider } from './ImageSlider';
