+++ src/components/MyPage/index.ts
@@ -1,0 +1,6 @@
+export { default as MyPageLayout } from './MyPageLayout';
+export { default as MyReviews } from './MyReviews';
+export { default as PageMenu } from './PageMenu';
+export { default as UpdatePassword } from './UpdatePassword';
+export { default as UpdateProfile } from './UpdateProfile';
+export { default as UserCard } from './UserCard';
