+++ src/components/common/Layout/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Header from 'components/Header';
-import Footer from 'components/Footer';
+import Header from 'components/common/Header';
+import Footer from 'components/common/Footer';
 import styled from '@emotion/styled';
 
 interface Props {
