+++ src/pages/myPage/index.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
-import MyPage from 'components/MyPage';
-import Loading from 'components/Loading';
-import MyReviews from 'components/MyPage/MyReviews';
+import { MyPageLayout, MyReviews } from 'components/mypage';
 import { MYPAGE_MENU_DEFAULT } from 'common/constant/string';
+import Loading from 'components/ui/Loading';
 import routes from 'common/constant/routes';
 import useUser from 'libs/useUser';
 
@@ -10,9 +9,9 @@ const myPage = (): React.ReactElement => {
   const { user } = useUser({ redirectTo: routes.LOGIN });
 
   return user && user.isLoggedIn ? (
-    <MyPage pageName={MYPAGE_MENU_DEFAULT} userInfo={user}>
+    <MyPageLayout pageName={MYPAGE_MENU_DEFAULT} userInfo={user}>
       <MyReviews />
-    </MyPage>
+    </MyPageLayout>
   ) : (
     <Loading />
   );
