+++ src/pages/search/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { useSearchData } from 'hooks';
-import { WriteButton, PostList, SinglePost } from 'components/Post';
+import { WriteButton, PostList, SinglePost } from 'components/post';
 import * as S from 'common/style/post';
-import Loading from 'components/Loading';
-import Modal from 'components/Modal';
+import Loading from 'components/ui/Loading';
+import Modal from 'components/ui/Modal';
 import useUser from 'libs/useUser';
 import routes from 'common/constant/routes';
 import { getReviewsMore, getReviewsFirst } from 'api/review';
