+++ src/pages/write.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import WritePost from 'components/Post/WritePost';
-import Loading from 'components/Loading';
+import { WritePost } from 'components/post';
+import Loading from 'components/ui/Loading';
 import routes from 'common/constant/routes';
 import useUser from 'libs/useUser';
